requirements.txt
xrpl-py==4.1.0
yfinance==0.2.54
# ecdsa==0.19.0
bech32==1.2.0
# bip_utils==2.9.3
bitcoin-utils==0.7.1
